+++ CarHubMobile/CarHub/src/main/java/com/worthwhilegames/carhubmobile/models/UserFuelRecord.java
@@ -48,9 +48,9 @@ public class UserFuelRecord extends UserBaseExpenseRecord {
     public FuelRecord toAPI() {
         FuelRecord toRet = new FuelRecord();
         toRet.setServerId(mRemoteId);
-        toRet.setVehicle(Integer.parseInt(mVehicle.getRemoteId()));
+        toRet.setVehicle(Long.parseLong(mVehicle.getRemoteId()));
         toRet.setDate(cDateFormat.format(new Date(mDate)));
-        toRet.setCategoryid(Integer.parseInt(CategoryRecord.fuelCategory(CategoryRecord.class).getRemoteId()));
+        toRet.setCategoryid(Long.parseLong(CategoryRecord.fuelCategory(CategoryRecord.class).getRemoteId()));
         toRet.setLocation(mLocation);
         toRet.setDescription(mDescription);
         toRet.setAmount((double) mAmount);
