+++ CarHubMobile/CarHub/src/main/java/com/worthwhilegames/carhubmobile/models/UserMaintenanceRecord.java
@@ -38,9 +38,9 @@ public class UserMaintenanceRecord extends UserBaseExpenseRecord {
     public MaintenanceRecord toAPI() {
         MaintenanceRecord toRet = new MaintenanceRecord();
         toRet.setServerId(mRemoteId);
-        toRet.setVehicle(Integer.parseInt(mVehicle.getRemoteId()));
+        toRet.setVehicle(Long.parseLong(mVehicle.getRemoteId()));
         toRet.setDate(cDateFormat.format(new Date(mDate)));
-        toRet.setCategoryid(Integer.parseInt(mCategory.getRemoteId()));
+        toRet.setCategoryid(Long.parseLong(mCategory.getRemoteId()));
         toRet.setLocation(mLocation);
         toRet.setDescription(mDescription);
         toRet.setAmount((double) mAmount);
